+++ src/statement/public.rs
@@ -36,6 +36,7 @@ use crate::client::ClientRef;
 use crate::cursor::Cursor;
 use crate::db_value::ToDbValue;
 use crate::error::Error;
+use crate::exec_result::ExecBatchResult;
 use crate::exec_result::ExecResult;
 use crate::row::Row;
 use crate::statement::StatementHolder;
@@ -91,7 +92,7 @@ impl<'sql> Statement<'sql> {
     ) -> Result<ExecResult, Error> {
         let mut holder = self.holder()?;
         let mut response = holder.execute(params)?;
-        Ok(ExecResult::new(holder.out_metadata(), &mut response))
+        Ok(ExecResult::new(holder.statement(), &mut response))
     }
 
     /// Executes a SQL statement against the database multiple times in one
@@ -99,10 +100,15 @@ impl<'sql> Statement<'sql> {
     pub fn execute_batch(
         &self,
         params: BindParameters,
-    ) -> Result<ExecResult, Error> {
+    ) -> Result<ExecBatchResult, Error> {
+        let num_execs = params.num_rows();
         let mut holder = self.holder()?;
         let mut response = holder.execute_batch(params)?;
-        Ok(ExecResult::new(holder.out_metadata(), &mut response))
+        Ok(ExecBatchResult::new(
+            holder.statement(),
+            num_execs,
+            &mut response,
+        ))
     }
 
     /// Executes the statement with the given named parameters and returns an
@@ -114,7 +120,7 @@ impl<'sql> Statement<'sql> {
     ) -> Result<ExecResult, Error> {
         let mut holder = self.holder()?;
         let mut response = holder.execute_named(params)?;
-        Ok(ExecResult::new(holder.out_metadata(), &mut response))
+        Ok(ExecResult::new(holder.statement(), &mut response))
     }
 
     /// Specifies that this statement should not be cached.